// src/models/TaskStateModel.ts
import type { TaskModel } from "./TaskModel";

// Estado -> Componente -> Filhos


export type TaskStateModel = {
  tasks: TaskModel[]; // histórico
  secondsRemaining: number;  // Home, CountDown, Histórico, MainForn,Button
  formattedSecondsRemaining: string; // Título, CountDown
  activeTask: TaskModel | null; // CountDown, Histórico, MainForm, Button
  currentCycle: number; // Home
  config: {
    workTime: number; // MainForm
    shortBreakTime: number; // MainForm
    longBreakTime: number; // MainForm
  };
};